Refactored vnr xscope python host example.
